pyproject.toml
[tool.poetry]
name = "balinese"
version = "0.1.0"
description = "Declare abyssinian infrastructure viw AWS CDK"
authors = ["oswalth <oswalth2@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/oswalth/balinese"
repository = "https://github.com/oswalth/balinese"
packages = [
  {include = "balinese", from = "./src/"}
]

[tool.poetry.dependencies]
python = "^3.10"
"aws-cdk.core" = "^1.163.2"
pydantic = "^1.9.1"
"aws-cdk.aws-ec2" = "^1.163.2"
"aws-cdk.aws-ecs" = "^1.163.2"
"aws-cdk.aws-ecs-patterns" = "^1.163.2"
"aws-cdk.aws-elasticloadbalancingv2" = "^1.163.2"
"aws-cdk.aws-ecr" = "^1.163.2"
"aws-cdk.aws-iam" = "^1.163.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
